+++ src/lib/py/subcommands.py
@@ -1,11 +1,19 @@
+import AddContaminants
+import CombineFasta
+import ComputeFDR
 import Convert2MzML
 import DownloadData
 import PhilosopherDatabase
+import PhilosopherFilter
 import RunFragPipe
 
 SUBCOMMANDS = [
+    AddContaminants.AddContaminants,
+    CombineFasta.CombineFasta,
+    ComputeFDR.ComputeFDR,
     Convert2MzML.Convert2MzML,
     DownloadData.DownloadData,
     PhilosopherDatabase.PhilosopherDatabase,
+    PhilosopherFilter.PhilosopherFilter,
     RunFragPipe.RunFragPipe,
 ]